refactor: centralize auth handling
