+++ src/controllers/emailServers.controller.ts
@@ -6,41 +6,7 @@ import {
 } from "../DALC/emailServers.dalc";
 import { empresa_getById_DALC } from "../DALC/empresas.dalc";
 import nodemailer from "nodemailer";
-
-// Función auxiliar para manejar la autenticación
-const handleAuth = (req: Request) => {
-    const authHeader = req.headers.authorization;
-    
-    // Si hay header de autenticación básica
-    if (authHeader && authHeader.startsWith('Basic ')) {
-        const base64Credentials = authHeader.split(' ')[1];
-        const credentials = Buffer.from(base64Credentials, 'base64').toString('ascii');
-        const [usernameBasic, password] = credentials.split(':');
-
-        // Credenciales permitidas
-        const validCredentials = [
-            ['A54APIDev', 'A54API4470Dev!'],
-            ['A54APIProd', 'A54API4470Prod!'],
-            ['A54APIProdUniversal', 'A54API4470ProdUniversal!']
-        ];
-
-        const isValid = validCredentials.some(([user, pass]) => user === usernameBasic && pass === password);
-
-        if (isValid) {
-            return { idEmpresa: 259, username: 'sistema' };
-        }
-        throw new Error('Credenciales inválidas');
-    }
-    // Si hay usuario autenticado vía JWT
-    else if (req.user) {
-        return { 
-            idEmpresa: req.user.idEmpresa, 
-            username: req.user.username 
-        };
-    }
-    
-    throw new Error('Se requiere autenticación');
-};
+import { handleAuth } from "../helpers/auth";
 
 export const getByEmpresa = async (req: Request, res: Response): Promise<Response> => {
     try {
